docs: :memo: Renamed simple recipe slightly with docs update
