+++ docs/src/pages/guide/index.mdx
@@ -54,7 +54,7 @@ Now, follow the next [usage](#usage) section. Also explore the [recipes] page to
 
 ## Usage 🚀
 
-Add the iles executable to the `scripts` section of your `package.json` if it's not already there.
+Add the iles executables to the `scripts` section of your `package.json` if it's not already there.
 
 ```json
   "scripts": {
