+++ playground/a-simple-iles-app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-simple-iles-app",
+  "name": "a-simple-iles-app",
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
