--- playground/my-simple-iles-app/src/app.ts
@@ -1,1 +1,0 @@
-export default {}
